Separate UplinkLib module into own file
